requirements.txt
streamlit==1.30.0
python-dotenv==1.0.1
google-generativeai==0.3.2
fpdf==1.7.2
markdown==3.5.2
SpeechRecognition==3.10.0
pyaudio==0.2.13
